design: FAQ 색상 변경
